refactor(app): extract BoatSetup to runtime/boat_setup.py (#72)
Move the 18-method boat-profile / specs / gains cluster out of Runtime
into a new BoatSetup collaborator (src/vanchor/runtime/boat_setup.py).
BoatSetup owns _boat_gains and _boat_gains_path; all other shared state
(config, state, bus, boats, controller, simulator) stays on Runtime and
is accessed via self._rt.  Runtime grows one-line shims for every
externally-called method so ui/server.py and the test suite are
unchanged.  Behaviour-preserving only: no logic changed.
